Split forms in components
